# Refactor io management