Make datalog and index to work with `io.Reader` and `io.Writer`, move file management to `api` module (prepare to make it cli), move datalog cursor management to index, speed up index rebuild a bit.