@eiri 2019-04-13
Bench:
```
benchmark                   old ns/op      new ns/op      delta
BenchmarkRebuildIndex-4     3150392135     2763142233     -12.29%
BenchmarkIndexOpen-4        85081828       88154305       +3.61%
BenchmarkOpen-4             84345943       90929238       +7.81%
BenchmarkWrite-4            25415          24704          -2.80%
BenchmarkRead-4             3930           4210           +7.12%

benchmark                   old allocs     new allocs     delta
BenchmarkRebuildIndex-4     245457         245557         +0.04%
BenchmarkIndexOpen-4        9602           9647           +0.47%
BenchmarkOpen-4             9645           9656           +0.11%
BenchmarkWrite-4            2              2              +0.00%
BenchmarkRead-4             1              1              +0.00%

benchmark                   old bytes     new bytes     delta
BenchmarkRebuildIndex-4     48589056      48632288      +0.09%
BenchmarkIndexOpen-4        41074582      41089038      +0.04%
BenchmarkOpen-4             41085526      41088130      +0.01%
BenchmarkWrite-4            1292          1291          -0.08%
BenchmarkRead-4             16            16            +0.00%
```